Switched app.js and server.js to fix loading order
