+++ package.json
@@ -2,7 +2,7 @@
   "name": "lobsterpages",
   "version": "1.0.0",
   "description": "Blog style web page for the team Almighty Lobsters",
-  "main": "app.js",
+  "main": "server.js",
   "scripts": {
     "start": "cross-env NODE_ENV=production node app.js",
     "build": "cross-env NODE_ENV=production webpack --progress --colors"
